requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cssselect2==0.7.0
Django==4.2.16
django-admin-interface==0.28.9
django-colorfield==0.11.0
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.54.1
html5lib==1.1
idna==3.10
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyphen==0.16.0
python-barcode==0.15.1
python-slugify==8.0.4
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tinycss2==1.3.0
typing_extensions==4.12.2
urllib3==2.2.3
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
